+++ python/raft-dask/raft_dask/test/conftest.py
@@ -4,9 +4,8 @@ import os
 
 import pytest
 
+from dask.distributed import Client
 from dask_cuda import LocalCUDACluster
-
-from raft_dask.common.utils import create_client
 
 os.environ["UCX_LOG_LEVEL"] = "error"
 
@@ -47,3 +46,26 @@ def ucx_client(ucx_cluster):
     client = create_client(ucx_cluster)
     yield client
     client.close()
+
+
+def create_client(cluster):
+    """
+    Create a Dask distributed client for a specified cluster.
+
+    Parameters
+    ----------
+    cluster : LocalCUDACluster instance or str
+        If a LocalCUDACluster instance is provided, a client will be created
+        for it directly. If a string is provided, it should specify the path to
+        a Dask scheduler file. A client will then be created for the cluster
+        referenced by this scheduler file.
+
+    Returns
+    -------
+    dask.distributed.Client
+        A client connected to the specified cluster.
+    """
+    if isinstance(cluster, LocalCUDACluster):
+        return Client(cluster)
+    else:
+        return Client(scheduler_file=cluster)
